+++ server.js
@@ -12,6 +12,7 @@ import './config/database.js'
 import { router as profilesRouter } from './routes/profiles.js'
 import { router as authRouter } from './routes/auth.js'
 import { router as postRouter } from './routes/posts.js'
+import { router as messageRouter } from './routes/messages.js'
 
 
 // create the express app
@@ -27,6 +28,7 @@ app.use(formData.parse())
 app.use('/api/profiles', profilesRouter)
 app.use('/api/auth', authRouter)
 app.use('/api/posts', postRouter)
+app.use('/api/messages', messageRouter)
 
 // handle 404 errors
 app.use(function (req, res, next) {
